requirements_v4.txt
# Newsletter API v4 Requirements
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data collection
feedparser==6.0.10
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2
trafilatura==1.6.4

# Video processing
youtube-transcript-api==0.6.1

# AI and scoring
openai==1.3.7
numpy==1.25.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
chardet==5.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring and logging
structlog==23.2.0

# Additional utilities
validators==0.22.0
urllib3==2.1.0
certifi==2023.11.17
